Minimize unnecessary usize/u64 casts across codebase

+++ kernel/src/memory/errors.rs
@@ -9,7 +9,7 @@ pub enum MemoryError {
     PhysicalToVirtual { addr: u64 },
 
     #[error("invalid page count: {pages}")]
-    InvalidPageCount { pages: u64 },
+    InvalidPageCount { pages: usize },
 
     #[error("out of memory")]
     OutOfMemory,
